+++ db/migrate/20220421084638_remove_unique_from_name_in_providers.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class RemoveUniqueFromNameInProviders < ActiveRecord::Migration[6.1]
+  def change
+    remove_index :providers, :name, unique: true
+  end
+end