+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Please view this file on the master branch, on stable branches it's out of date.
 ### Fixed
 - Providers import (@kmarszalek)
 
+### Removed
+- Unique constraints from name and abbreviation for Provider (@kmarszalek)
+
 ## [3.36.0] 2022-04-19
 
 ### Changed